fix: fix seed file to devide env
